PythonBackend/requirements.txt
Flask>=2.3.2
Werkzeug>=2.3.6
flask-cors
flask-httpauth
langchain
langchain-community
openai==0.28
python-dotenv
PyPDF2
firebase-admin
redis
tiktoken
faiss-cpu